@@ -23,7 +23,7 @@ namespace WebApi
                 .AddAspNetInfrastructure();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             app.UseAspNetInfrastructure(env, Configuration);
         }
