@@ -1,22 +1,24 @@
-﻿
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Infrastructure;
 using Infrastructure.Logging;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 
 namespace WebApi
 {
     public class Program
     {
-        public static async Task Main(string[] args)
-        {
-            await WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
-                .ConfigureLoggingPlumbing()
-                .Build()
-                .RunAsync();
-        }
- 
+        public static Task Main(string[] args) => CreateHostBuilder(args).Build().RunAsync();
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder
+                        .UseStartup<Startup>()
+                        .ConfigureLoggingPlumbing()
+                        .CaptureStartupErrors(true);
+                });
     }
 }
